+++ integration/openjpa/src/main/java/com/blazebit/persistence/integration/openjpa/OpenJPAJpaProvider.java
@@ -319,6 +319,11 @@ public class OpenJPAJpaProvider implements JpaProvider {
     }
 
     @Override
+    public boolean supportsUpdateSetAssociationId() {
+        return true;
+    }
+
+    @Override
     public boolean supportsTransientEntityAsParameter() {
         return true;
     }